fix(engine): the callback instrument stops perturbing what it measures
Review of #724. Ten findings; the blocking ones all came down to the
instrument adding cost or noise to the system it exists to observe.

App-side rebuilds (the blocking one): the telemetry is off the
render-rate path entirely. It no longer rides EngineSnapshot or
EngineStatus — a counter that ticks on every audio callback made every
projected LooperState unequal to the last, permanently defeating
LooperRepository's dedupe gate so an idle rig re-broadcast its whole
state on every poll tick, and allocated four Lists per tick doing it.
It is now a pull: AudioEngine.callbackTelemetry() /
LooperRepository.callbackTelemetry. EngineStatus is doc-only against
master; EngineSnapshot is purely additive.

miniaudio hook: the -EPIPE sites sit inside a retry loop, so a wedged
device turned one physical dropout into thousands of counted xruns and
thousands of RT-thread calls. Each read/write call now reports once.
The hook install is refcounted, so engine A closing its device no
longer silences counting for engine B's open one — which would have
put xrun_count back to a permanent 0, the exact symptom #722 exists to
end. And the build gate is real: engine_telemetry_gate.h is shared
with miniaudio_impl.c, so -DLE_CALLBACK_TELEMETRY=0 removes the hook
call sites and the install, not just the accumulators.

Measurement correctness: the deadline now comes from each callback's
own frame count, not buffer_frames — the duplex loop delivers
min(capture, playback) chunks and short reads, and judging a fraction
of a period against a whole one reported "all clear" on a struggling
device. A gap that a counted recovery already explains is suppressed,
so gap_events is independent of xruns instead of double-counting one
dropout under two names. Buckets and xruns widened to 64-bit (a 32-bit
bucket wraps in ~33 days of uptime). The disarm summary is emitted on
the stalled-handshake path too, marked stalled=1 — that path is
exactly when the numbers matter most. The telemetry seed no longer
reads a stale a_buffer_frames, and clears a_xruns in lockstep with the
windows it summarises. The single-writer claim is scoped to the fields
that actually have one writer.

+++ packages/segno_engine/src/core/segno_engine_api.h
@@ -669,8 +669,13 @@ typedef struct le_track_snapshot {
  * le_engine_process between two monotonic clock reads (engine_miniaudio.c), so a
  * duration here is what the DEVICE's callback thread actually spent, and an
  * entry-to-entry gap is what the DEVICE actually experienced between periods.
- * The deadline it is compared against is the period budget — buffer_frames /
- * sample_rate — reported as cb_budget_us.
+ * The deadline it is compared against is that callback's OWN budget — its frame
+ * count over the sample rate — NOT buffer_frames / sample_rate. The two differ
+ * on exactly the path this exists to measure: miniaudio's duplex loop drives the
+ * callback in min(capture, playback) chunks, split further by the converter and
+ * by short readi() returns, so a callback routinely handles a fraction of a
+ * playback period and judging it against a whole one would report "all clear" on
+ * a struggling device. cb_budget_us reports the most recent callback's budget.
  *
  * TWO WINDOWS, because the bug being hunted (#722: clicks ONLY while a
  * performance capture is armed) is a comparison, not an absolute: `cb_session`
@@ -704,23 +709,32 @@ typedef enum le_xrun_kind {
 /* One accumulation window of callback telemetry. Every counter is monotonic
  * within its window and is only ever cleared by the event that owns the window
  * (a fresh configure/start for the session window; le_perf_arm for the armed
- * one). Durations are microseconds: a nanosecond figure would overflow uint32
- * after 4.3 s and nothing here is finer than a microsecond anyway. */
+ * one). Counts are 64-bit throughout — a 32-bit histogram bucket at ~1500
+ * callbacks/second wraps inside a month, and this has to stay readable on an
+ * appliance that has been up since the last OTA. Durations are microseconds: a
+ * nanosecond figure would overflow uint32 after 4.3 s and nothing here is finer
+ * than a microsecond anyway. */
 typedef struct le_cb_window_snapshot {
   uint64_t calls;      /* device callbacks observed in this window */
   uint64_t late_calls; /* of those, ones whose duration exceeded the budget */
-  /* Entry-to-entry gaps longer than 1.5 periods: the DERIVED starvation signal.
-   * The device handing us a period more than half a period late means it
+  /* Entry-to-entry gaps longer than 1.5 of the previous callback's budget: the
+   * DERIVED starvation signal. The device coming back that late means it
    * starved, whether or not the backend admitted an xrun — this is what works
-   * on CoreAudio, where miniaudio exposes no xrun signal at all. */
+   * on CoreAudio, where miniaudio exposes no xrun signal at all.
+   *
+   * INDEPENDENT OF `xruns` by construction: an ALSA -EPIPE recovery also
+   * produces a long gap, so the callback that follows a counted recovery has
+   * its gap suppressed. One physical dropout is counted once, under one name,
+   * and max_gap_us is never pinned by a recovery stall. A non-zero gap_events
+   * therefore means starvation the backend never reported. */
   uint64_t gap_events;
   uint32_t max_us;     /* worst callback duration seen */
   uint32_t mean_us;    /* mean callback duration over `calls` */
-  uint32_t max_gap_us; /* worst entry-to-entry gap (0 when none exceeded) */
-  uint32_t buckets[LE_CB_BUCKETS]; /* duration histogram; see LE_CB_BUCKETS */
+  uint32_t max_gap_us; /* worst counted gap (0 when none exceeded) */
+  uint64_t buckets[LE_CB_BUCKETS]; /* duration histogram; see LE_CB_BUCKETS */
   /* Real backend dropouts, indexed by le_xrun_kind. Their sum over the session
    * window is exactly what xrun_count reports. */
-  uint32_t xruns[LE_XRUN_KINDS];
+  uint64_t xruns[LE_XRUN_KINDS];
 } le_cb_window_snapshot;
 
 /* Lock-free snapshot of engine state, published by the audio thread and read by
@@ -910,9 +924,12 @@ typedef struct le_snapshot {
    * an engine whose device never ran — the native test pump processes blocks
    * without a device and therefore has no deadline to miss. */
 
-  /* The period deadline every duration below is judged against, in
-   * microseconds: buffer_frames / sample_rate. 0 = no device has been opened,
-   * which is also the "telemetry inert" state (nothing is accumulated). */
+  /* The deadline the MOST RECENT callback was judged against, in microseconds:
+   * that callback's frame count over the sample rate. Usually constant, but it
+   * tracks a duplex loop that varies the block size rather than pretending
+   * buffer_frames is the answer (see the note above). 0 = no device has been
+   * opened, or none of its callbacks has landed yet — which is also the
+   * "telemetry inert" state, in which nothing is accumulated at all. */
   uint32_t cb_budget_us;
   le_cb_window_snapshot cb_session; /* since the device started */
   le_cb_window_snapshot cb_armed;   /* since the most recent le_perf_arm */
